requirements.txt
-e .
aws-cdk.aws-iot
boto3
botostubs